BSKY_validation/requirements.txt
atproto
python-dotenv
requests

